Library to access Microchip UNIO devices on Nanode
All Nanodes (should!) have a Microchip 11AA02E48 eeprom device connected
to pin 7 of the microcontroller.  (This is the three-pin surface mount
device on the back of the board.)  This device holds the Nanode's unique
MAC address.  There's plenty of free space in it, though; the MAC
address takes up 6 bytes out of 256, and this library enables you to use
the device for storage.